+++ src/plugins/intel_npu/src/compiler_adapter/include/driver_graph.hpp
@@ -21,7 +21,7 @@ public:
                 ze_graph_handle_t graphHandle,
                 NetworkMetadata metadata,
                 const Config& config,
-                std::optional<std::unique_ptr<BlobContainer>> blob);
+                std::unique_ptr<BlobContainer> blob);
 
     size_t export_blob(std::ostream& stream) const override;
