+++ src/plugins/intel_npu/src/compiler_adapter/src/plugin_graph.cpp
@@ -17,10 +17,7 @@ PluginGraph::PluginGraph(const std::shared_ptr<ZeGraphExtWrappers>& zeGraphExt,
                          NetworkMetadata metadata,
                          std::unique_ptr<BlobContainer> blobPtr,
                          const Config& config)
-    : IGraph(graphHandle,
-             std::move(metadata),
-             config,
-             std::optional<std::unique_ptr<BlobContainer>>(std::move(blobPtr))),
+    : IGraph(graphHandle, std::move(metadata), config, std::move(blobPtr)),
       _zeGraphExt(zeGraphExt),
       _zeroInitStruct(zeroInitStruct),
       _compiler(compiler),
